requirements.txt
python-dotenv==1.0.1
Pillow==9.3.0
pandas==1.5.2
requests==2.31.0
tqdm==4.64.1
torch==2.2.2
torchmetrics==1.3.2
basemap==1.4.1
numpy==1.23.4